Preserve personal configuration /etc/default/fsser
Don't overwrite personal configurations on re-install/update
if this file already exists

+++ pcserver/daemon.sh
@@ -1,9 +1,13 @@
-#!/bin/sh
+#!/bin/bash
 #
 # Create a XD-2031 daemon
 #
 install -m 755 fsser.init.d.localized /etc/init.d/fsser
-install -m 755 fsser.default /etc/default/fsser
+if [ -f /etc/default/fsser ] ; then
+	echo "Personal configuration /etc/default/fsser preserved."
+else
+	install -m 755 fsser.default /etc/default/fsser
+fi
 update-rc.d fsser defaults
 service fsser stop
 service fsser start
